add support to mapTiler

+++ src/utils/getHeightMap.ts
@@ -6,24 +6,33 @@ type T = {
   heightmapImage?: ImageBitmap
 }
 
-const getHeightmapData = (mapType: MapType, settings: Settings, token: string, isDebug: boolean, resolution: number) => {
+const getHeightmapData = (mapType: MapType, settings: Settings, isDebug: boolean, resolution: number) => {
   const data: GenerateMapOption = {
     mapType,
     settings,
-    token,
     isDebug,
     resolution,
   }
   const generateMapOption = JSON.parse(JSON.stringify(data))
   const worker = new GetHeightMapWorker()
+
   const result = new Promise<T>((resolve) => {
-    worker.addEventListener('message', (e) => {
+    const handleMessage = (e: MessageEvent) => {
       if (e.data) {
-        resolve(e.data)
-        worker.terminate()
+        if (e.data.type === 'total') {
+          useEvent('tile:total', e.data.number)
+        } else if (e.data.type === 'progress') {
+          useEvent('tile:progress')
+        } else {
+          resolve(e.data)
+          worker.removeEventListener('message', handleMessage)
+          worker.terminate()
+        }
       }
-    }, { once: true })
+    }
+    worker.addEventListener('message', handleMessage)
   })
+
   worker.postMessage(generateMapOption)
   return result
 }
@@ -35,19 +44,11 @@ const getHeightmapData = (mapType: MapType, settings: Settings, token: string, i
  * @param resolution for preview
  * @return Promise\<Float32Array, ImageBitmap\>
  */
-export const getHeightmap = async (mapType: MapType = 'cs1', preview = false, isDebug = false, resolution?: number) => {
+export const getHeightmap = async (mapType: MapType = 'cs1', isDebug = false, resolution?: number) => {
   try {
     const { settings } = useMapbox().value
-    const config = useRuntimeConfig()
-    if (!(preview === true || mapType === 'cs1') && !isTokenValid()) {
-      throw new Error('Invaid access token')
-    }
-    let token = (settings.gridInfo === 'cs1' || preview === true) ? (settings.accessToken || config.public.mapboxToken) : settings.accessToken
-    if (mapType === 'ocean') {
-      token = config.public.maptilerToken
-    }
     const _resolution = resolution || settings.resolution
-    const result = await getHeightmapData(mapType, settings, token, isDebug, _resolution)
+    const result = await getHeightmapData(mapType, settings, isDebug, _resolution)
     return result
   } catch (error) {
     console.error('An error occurred in getHeightMap:', error)
